+++ src/parser/unit.h
@@ -11,14 +11,14 @@ public:
   Unit(std::string unitString);
   ~Unit();
 
-  float evalUnit(float x);
+  double evalUnit(double x);
   std::string getUnitString() { return uString; }
 
 private:
-  void performOps(std::vector<std::variant<Operation, float>> &transformed,
+  void performOps(std::vector<std::variant<Operation, double>> &transformed,
                   char op1, char op2);
-  std::vector<std::variant<Operation, float>>
-  transformUnit(std::vector<std::variant<Operation, float>> parts);
-  std::vector<std::variant<Unit, Operation, float>> parts;
+  std::vector<std::variant<Operation, double>>
+  transformUnit(std::vector<std::variant<Operation, double>> parts);
+  std::vector<std::variant<Unit, Operation, double>> parts;
   std::string uString;
 };
